+++ test/firestore_test.dart
@@ -396,15 +396,15 @@ void main() {
         });
       });
 
-      test('getCollections', () async {
-        var doc = app.firestore().document('tests/get_collections');
+      test('listCollections', () async {
+        var doc = app.firestore().document('tests/list_collections');
         // Create an element in a sub collection to make sure the collection
         // exists
         await doc
             .collection('sub')
             .document('item')
             .setData(DocumentData.fromMap({}));
-        var collections = await doc.getCollections();
+        var collections = await doc.listCollections();
         expect(collections.any((CollectionReference col) => col.id == 'sub'),
             isTrue);
       });
@@ -476,10 +476,10 @@ void main() {
         expect(doc.path, 'users/abc');
       });
 
-      test('getCollections', () async {
+      test('listCollections', () async {
         // create a document to make sure the collection is created
         await ref.document('any').setData(DocumentData.fromMap({}));
-        var collections = await app.firestore().getCollections();
+        var collections = await app.firestore().listCollections();
         // Find our collection
         expect(
           collections.any((CollectionReference col) => col.path == ref.path),
@@ -967,7 +967,7 @@ void main() {
         });
 
         var error = await result.catchError((error) => error);
-        expect(error.toString(), contains('FAILED_PRECONDITION'));
+        expect(error.toString(), contains('does not match the required base version'));
 
         expect((await doc1Ref.get()).data.toMap(), {'value': 10});
         expect((await doc2Ref.get()).data.toMap(), {'value': 20});
@@ -1058,8 +1058,8 @@ void main() {
 
     group('Collection Groups', () {
       test('fetch documents in a collection group', () async {
-        var doc1 = app.firestore().document('tests/group_a/doc1');
-        var doc2 = app.firestore().document('tests/nested/group_a/doc2');
+        var doc1 = app.firestore().document('tests/one/group_a/doc1');
+        var doc2 = app.firestore().document('tests/two/group_a/doc2');
         await doc1.setData(DocumentData.fromMap({'value': 1}));
         await doc2.setData(DocumentData.fromMap({'value': 2}));